# legalai-backend/app/static/openapi.yaml
openapi: 3.0.3
info:
  title: Legal Lawyer AI API
  version: "1.0.0"
  description: Swagger spec generated from the provided Postman collection.
servers:
  - url: http://localhost:5000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
...
security: []  # default none; per-route add as needed
paths:
...
  /api/v1/pathways:
    get:
      tags: [Pathways]
      summary: List pathways
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: language
          schema: { type: string }
      responses:
        "200": { description: Success }
    post:
      tags: [Pathways]
      summary: Create pathway (Admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              title: Divorce Process in Pakistan
              summary: Step-by-step guide for divorce proceedings
              steps:
                - step: 1
                  title: File petition
                  description: File in family court
                - step: 2
                  title: Notice period
                  description: 90 days waiting period
              category: family
              language: en
              tags: [divorce, family-law]
      responses:
        "200": { description: Success }
  /api/v1/pathways/{pathwayId}:
    get:
      tags: [Pathways]
      summary: Get single pathway
      parameters:
        - in: path
          name: pathwayId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success }
    put:
      tags: [Pathways]
      summary: Update pathway (Admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: pathwayId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
      responses:
        "200": { description: Success }
    delete:
      tags: [Pathways]
      summary: Delete pathway (Admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: pathwayId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success }

  # -------------------- CHECKLISTS --------------------
  /api/v1/checklists/categories:
    get:
      tags: [Checklists]
      summary: List checklist categories
      responses:
        "200": { description: Success }
    post:
      tags: [Checklists]
      summary: Create category (Admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              title: Court Documents
              icon: "📄"
              order: 1
      responses:
        "200": { description: Success }
  /api/v1/checklists/categories/{categoryId}:
    put:
      tags: [Checklists]
      summary: Update category (Admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: categoryId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
      responses:
        "200": { description: Success }
    delete:
      tags: [Checklists]
      summary: Delete category (Admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: categoryId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success }
  /api/v1/checklists/items:
    get:
      tags: [Checklists]
      summary: List checklist items
      parameters:
        - in: query
          name: categoryId
          required: false
          schema: { type: integer }
      responses:
        "200": { description: Success }
    post:
      tags: [Checklists]
      summary: Create item (Admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              categoryId: 1
              text: CNIC Copy
              required: true
              order: 1
      responses:
        "200": { description: Success }
  /api/v1/checklists/items/{itemId}:
    put:
      tags: [Checklists]
      summary: Update item (Admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: itemId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
      responses:
        "200": { description: Success }
    delete:
      tags: [Checklists]
      summary: Delete item (Admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: itemId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success }

  # -------------------- DRAFTS --------------------
  /api/v1/drafts/generate:
    post:
      tags: [Drafts]
      summary: Generate draft
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              templateId: 1
              answers:
                landlordName: Ali Ahmed
                tenantName: Usman Khan
                rent: "50000"
                address: DHA Phase 5, Lahore
              userSnapshot:
                name: Ahmed Khan
                cnic: "12345-1234567-1"
                phone: "03001234567"
      responses:
        "200": { description: Success }
  /api/v1/drafts:
    get:
      tags: [Drafts]
      summary: List drafts
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Success }
  /api/v1/drafts/{draftId}:
    get:
      tags: [Drafts]
      summary: Get single draft
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: draftId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success }
    delete:
      tags: [Drafts]
      summary: Delete draft
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: draftId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success }
  /api/v1/drafts/{draftId}/export:
    get:
      tags: [Drafts]
      summary: Export draft
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: draftId
          required: true
          schema: { type: integer }
        - in: query
          name: format
          required: true
          schema:
            type: string
            enum: [txt, pdf, docx]
      responses:
        "200":
          description: File stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

  # -------------------- CHAT --------------------
...
  /api/v1/reminders:
    post:
      tags: [Reminders]
      summary: Create reminder
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              title: Court Hearing
              notes: Family court hearing at 10 AM
              scheduledAt: "2025-01-20T10:00:00Z"
              timezone: Asia/Karachi
      responses:
        "200": { description: Success }
    get:
      tags: [Reminders]
      summary: List reminders
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Success }
  /api/v1/reminders/{reminderId}:
    put:
      tags: [Reminders]
      summary: Update reminder
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: reminderId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              title: Updated Court Hearing
              isDone: true
      responses:
        "200": { description: Success }
    delete:
      tags: [Reminders]
      summary: Delete reminder
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: reminderId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success }
  /api/v1/reminders/register-device-token:
    post:
      tags: [Reminders]
      summary: Register device token
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              platform: android
              token: fcm_device_token_here_very_long_string
      responses:
        "200": { description: Success }

  # -------------------- ADMIN KNOWLEDGE --------------------
...